+++ src/main/javascript/Package/index.js
@@ -1,9 +1,15 @@
 'use strict';
 
 const path = require('path');
+
 const Validator = require('./Validator');
+const PackageBuilder = require('./PackageBuilder');
+const convertAppName = require('./convertAppName');
+
 
 exports = module.exports =
 {
-  Validator: Validator
+  Validator: Validator,
+  PackageBuilder: PackageBuilder,
+  convertAppName: convertAppName
 };