+++ src/main/javascript/command/dpat-package.js
@@ -4,10 +4,9 @@ const program = require("commander");
 const archiver = require("archiver");
 const fs = require("fs");
 
-const defaultWebpackConfig = require.resolve('../webpack/webpack.config-distribution');
 const __path = require("path");
 const project = require("../Project");
-const packageFilename = 'app.zip';
+const PackageBuilder = require("../Package").PackageBuilder;
 
 /**
  * @param {String} path
@@ -26,7 +25,7 @@ function action(path, cmd)
 
   let packagePath;
   try {
-    packagePath = dpProject.runPackage(projectDir, packageFilename);
+    packagePath = dpProject.runPackage(projectDir, new PackageBuilder());
   } catch (e) {
     console.log(`ERROR: failed to package the distribution folder: ${projectDir}`);
     process.exit(1);